requirements.txt
fastapi
uvicorn
python-dotenv
openai
reportlab
python-multipart
pypdf
pytest
pytest-env
python-docx
